postgresql 8.3 changes
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@1927 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/yourpayprocessor.cpp
@@ -272,7 +272,7 @@ int  YourPayProcessor::doAuthorize(const int pccardid, const int pcvv, const dou
 	      "       AS next_seq "
 	      "  FROM ccpay "
 	      " WHERE (ccpay_order_number=:ccpay_order_number);");
-  ypq.bindValue(":ccpay_order_number", pneworder.toInt());
+  ypq.bindValue(":ccpay_order_number", pneworder);
   ypq.exec();
   if (ypq.first() && ! ypq.value("next_seq").isNull())
     oidstr = oidstr + "-" + ypq.value("next_seq").toString();
